third_party/WebKit/LayoutTests/fast/canvas/canvas-drawImage-live-video-expected.txt

Verify that consecutive drawImage from a live video correctly propagates frame updates.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS imagesAreTheSame is false
